load_to_bigquery/requirements.txt
gunicorn
flask
requests
beautifulsoup4
google-cloud-storage
google-cloud-bigquery

